refactor: add check user access to medcard

+++ api/medical_record/allergy.py
@@ -8,6 +8,7 @@ from models.medical_record.allergy import Allergy, AllergyPK, AllergyCreate, All
 from models.user import User
 from services.medical_record.allergy import AllergyService
 from services.auth import get_current_user
+from services.user import check_user_access_to_medcard
 
 
 router = APIRouter(
@@ -20,7 +21,7 @@ router = APIRouter(
 async def get_allergies_by_medcard_num(medcard_num: int,
                                        user: User = Depends(get_current_user),
                                        service: AllergyService = Depends()):
-    if user:
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
         return service.get_allergies_by_medcard_num(medcard_num=medcard_num)
     else:
         raise HTTPException(
@@ -29,28 +30,52 @@ async def get_allergies_by_medcard_num(medcard_num: int,
 
 
 @router.post('/one', response_model=Allergy)
-async def get_allergy_by_pk(allergy_pk: AllergyPK,
+async def get_allergy_by_pk(medcard_num: int,
+                            allergy_pk: AllergyPK,
                             user: User = Depends(get_current_user),
                             service: AllergyService = Depends()):
-    return service.get_allergy_by_pk(allergy_pk=allergy_pk)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.get_allergy_by_pk(allergy_pk=allergy_pk)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.post('/', response_model=Allergy)
 async def add_allergy(allergy_data: AllergyCreate,
-                      user: User = Depends(get_current_user),
+                      medcard_num: int,
+user: User = Depends(get_current_user),
                       service: AllergyService = Depends()):
-    return service.add_new_allergy(allergy_data=allergy_data)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.add_new_allergy(allergy_data=allergy_data)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.put('/', response_model=Allergy)
 async def update_allergy(allergy_data: AllergyUpdate,
-                         user: User = Depends(get_current_user),
+                         medcard_num: int,
+user: User = Depends(get_current_user),
                          service: AllergyService = Depends()):
-    return service.update_allergy(allergy_data=allergy_data)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.update_allergy(allergy_data=allergy_data)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.delete('/')
 async def delete_allergy(allergy_pk: AllergyPK,
-                         user: User = Depends(get_current_user),
+                         medcard_num: int,
+user: User = Depends(get_current_user),
                          service: AllergyService = Depends()):
-    service.delete_allergy(allergy_pk=allergy_pk)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        service.delete_allergy(allergy_pk=allergy_pk)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
