+++ api/medical_record/extra_class.py
@@ -1,6 +1,8 @@
 from fastapi import (
     APIRouter,
     Depends,
+    HTTPException,
+    status,
 )
 from models.medical_record.extra_class import (
     ExtraClassPK,
@@ -11,6 +13,7 @@ from models.medical_record.extra_class import (
 from models.user import User
 from services.medical_record.extra_class import ExtraClassService
 from services.auth import get_current_user
+from services.user import check_user_access_to_medcard
 
 
 router = APIRouter(
@@ -23,32 +26,61 @@ router = APIRouter(
 async def get_extra_cas_by_medcard_num(medcard_num: int,
                                        user: User = Depends(get_current_user),
                                        service: ExtraClassService = Depends()):
-    return service.get_extra_classes_by_medcard_num(medcard_num=medcard_num)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.get_extra_classes_by_medcard_num(medcard_num=medcard_num)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.post('/one', response_model=ExtraClass)
 async def get_extra_class_by_pk(extra_class_pk: ExtraClassPK,
+                                medcard_num: int,
                                 user: User = Depends(get_current_user),
                                 service: ExtraClassService = Depends()):
-    return service.get_extra_class_by_pk(extra_class_pk=extra_class_pk)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.get_extra_class_by_pk(extra_class_pk=extra_class_pk)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.post('/', response_model=ExtraClass)
 async def add_extra_class(extra_class_data: ExtraClassCreate,
+                          medcard_num: int,
                           user: User = Depends(get_current_user),
                           service: ExtraClassService = Depends()):
-    return service.add_new_extra_class(extra_class_data=extra_class_data)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.add_new_extra_class(extra_class_data=extra_class_data)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.put('/', response_model=ExtraClass)
 async def update_extra_class(extra_class_data: ExtraClassUpdate,
+                             medcard_num: int,
                              user: User = Depends(get_current_user),
                              service: ExtraClassService = Depends()):
-    return service.update_extra_class(extra_class_data=extra_class_data)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.update_extra_class(extra_class_data=extra_class_data)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.delete('/')
 async def delete_extra_class(extra_class_pk: ExtraClassPK,
+                             medcard_num: int,
                              user: User = Depends(get_current_user),
                              service: ExtraClassService = Depends()):
-    service.delete_extra_class(extra_class_pk=extra_class_pk)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.delete_extra_class(extra_class_pk=extra_class_pk)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
