+++ api/medical_record/gg_injection.py
@@ -1,6 +1,8 @@
 from fastapi import (
     APIRouter,
     Depends,
+    HTTPException,
+    status,
 )
 from models.medical_record.gg_injection import (
     GammaGlobulinInjectionPK,
@@ -11,6 +13,7 @@ from models.medical_record.gg_injection import (
 from models.user import User
 from services.medical_record.gg_injection import GammaGlobulinInjectionService
 from services.auth import get_current_user
+from services.user import check_user_access_to_medcard
 
 
 router = APIRouter(
@@ -21,35 +24,63 @@ router = APIRouter(
 
 @router.get('/', response_model=list[GammaGlobulinInjection])
 async def get_gamma_globulin_injections_by_medcard_num(medcard_num: int,
-                                           user: User = Depends(
-                                               get_current_user),
+                                           user: User = Depends(get_current_user),
                                            service: GammaGlobulinInjectionService = Depends()):
-    return service.get_gamma_globulin_injections_by_medcard_num(medcard_num=medcard_num)
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.get_gamma_globulin_injections_by_medcard_num(medcard_num=medcard_num)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.post('/one', response_model=GammaGlobulinInjection)
 async def get_gamma_globulin_injection_by_pk(gg_injection_pk: GammaGlobulinInjectionPK,
-                                 user: User = Depends(get_current_user),
-                                 service: GammaGlobulinInjectionService = Depends()):
-    return service.get_gamma_globulin_injection_by_pk(gg_injection_pk=gg_injection_pk)
+                                             medcard_num: int,
+                                             user: User = Depends(get_current_user),
+                                             service: GammaGlobulinInjectionService = Depends()):
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.get_gamma_globulin_injection_by_pk(gg_injection_pk=gg_injection_pk)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.post('/', response_model=GammaGlobulinInjection)
 async def add_gamma_globulin_injection(gg_injection_data: GammaGlobulinInjectionCreate,
-                           user: User = Depends(get_current_user),
-                           service: GammaGlobulinInjectionService = Depends()):
-    return service.add_new_gamma_globulin_injection(gg_injection_data=gg_injection_data)
+                                       medcard_num: int,
+                                       user: User = Depends(get_current_user),
+                                       service: GammaGlobulinInjectionService = Depends()):
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.add_new_gamma_globulin_injection(gg_injection_data=gg_injection_data)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.put('/', response_model=GammaGlobulinInjection)
 async def update_gamma_globulin_injection(gg_injection_data: GammaGlobulinInjectionUpdate,
-                              user: User = Depends(get_current_user),
-                              service: GammaGlobulinInjectionService = Depends()):
-    return service.update_gamma_globulin_injection(gg_injection_data=gg_injection_data)
+                                          medcard_num: int,
+                                          user: User = Depends(get_current_user),
+                                          service: GammaGlobulinInjectionService = Depends()):
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.update_gamma_globulin_injection(gg_injection_data=gg_injection_data)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
 
 
 @router.delete('/')
 async def delete_gamma_globulin_injection(gg_injection_pk: GammaGlobulinInjectionPK,
-                              user: User = Depends(get_current_user),
-                              service: GammaGlobulinInjectionService = Depends()):
-    service.delete_gamma_globulin_injection(gg_injection_pk=gg_injection_pk)
+                                          medcard_num: int,
+                                          user: User = Depends(get_current_user),
+                                          service: GammaGlobulinInjectionService = Depends()):
+    if check_user_access_to_medcard(user=user, medcard_num=medcard_num):
+        return service.delete_gamma_globulin_injection(gg_injection_pk=gg_injection_pk)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN
+        )
